refactor: move hmac verification to a helper util
