+++ src/sdks/projects.ts
@@ -2,6 +2,7 @@ import type {API, Endpoints, Event, Id} from '../rest/index.ts';
 import {Request} from '../util/fetch.ts';
 import {sdk} from './create.ts';
 import type {PossibleWebhookIDs} from '../util/types.ts';
+import {verifyHmac} from '../index.ts';
 
 /**
  * Projects SDK client
@@ -105,33 +106,12 @@ export const projects = sdk(client => {
 		 * @param secret The secret provided upon webhook creation to verify the signature. (e.x: whsec_xxxxx)
 		 */
 		async constructEvent(body: string, signature: string, secret: string) {
-			const encoder = new TextEncoder();
-			const encodedBody = encoder.encode(body);
-
-			const key = await crypto.subtle.importKey(
-				'raw',
-				encoder.encode(secret),
-				{name: 'HMAC', hash: 'SHA-256'},
-				false,
-				['sign'],
-			);
-
-			const signatureBuffer = await crypto.subtle.sign(
-				'HMAC',
-				key,
-				encodedBody,
-			);
-
-			const finalSig = Array.from(new Uint8Array(signatureBuffer))
-				.map(byte => byte.toString(16).padStart(2, '0'))
-				.join('');
-
-			if (signature.toLowerCase() !== finalSig) {
+			const hmacVerified = await verifyHmac(body, signature, secret);
+			if (!hmacVerified) {
 				throw new Error('Invalid signature');
 			}
 
 			const event = JSON.parse(body) as Event;
-
 			return event;
 		},
 		async getAll(projectId?: Id<'project'>) {
